Check values of enum variants in test (#32)
For using a HashSet as a database the values of each enum variant are
ignored. Thus each enum variant has a stable hash value, which is not
changed by the stored value. This allows updating old data with new data
in the HashSet, but it non obvious behavior in other cases, like
checking the enum variant with value. Thus tests have been adapted until
a better design is found.